build: introduce DT_NAME property for devicetree supported devices
Every devicetree supported devices defines a 'compatible' property
which is used to specifies the "exact device". The 'compatible'
property contains a sorted list of strings starting with the exact
name of the machine, followed by an optional list of boards it is
compatible with sorted from most compatible to least.

The reasoning behind this scheme is the observation that in the majority
of cases, a single arch/driver can support a large number of boards
if they all use the same SoC, or same family of SoCs.  However,
invariably there will be some exceptions where a specific board will
require special setup code that is not useful in the generic case.
Special cases could be handled by explicitly checking for the
troublesome board(s) in generic setup code, but doing so very quickly
becomes ugly and/or unmaintainable if it is more than just a couple of
cases.

Instead, the compatible list allows a generic machine_desc to provide
support for a wide common set of boards by specifying "less
compatible" values in the 'compatible' property list.

<https://www.kernel.org/doc/Documentation/devicetree/usage-model.txt>
<http://elinux.org/Device_Tree_Usage#Understanding_the_compatible_Property>

Warning: Don't use wildcard compatible values, like "qcom,ipq806x"
or similar. Vendors and other users will invariably make a change
that breaks your wildcard assumptions the moment it is too late
to change it. Instead, choose a specific silicon implementations
and make all subsequent silicon compatible with it.

Signed-off-by: Christian Lamparter <chunkeey@googlemail.com>