apm821xx: make it possible to go into a poweroff-like state
The BookE watchdog can't be disabled by software. Once it has
been actived, the watchdog will need continous servicing in
order to prevent it from rebooting the device. The timeout
is prettly low (~6 secs) and this can cause problems for
users which want to power down their devices.

One option is to remove the BOOKE watchdog from the kernel
config.

Another option is to make use of a little-known "hardware
feature" that is specific to the MBL. The MBL can disable
the NOR-Chip on which U-Boot is stored. This will cause a
the reboot and prevent the HDDs from spinning up.

Note: This isn't perfect, the gpio-poweroff was never
intended to be used in this way... So it will fire off
a WARN_ON(1) since the GPIO pin doesn't actually power
off the device.

Signed-off-by: Christian Lamparter <chunkeey@googlemail.com>